godocs and restructure packages.
