+++ cmd/daemon.go
@@ -1,1 +1,3 @@
 package cmd
+
+// TODO: `testground daemon` command. Starts the testground daemon.
