+++ cmd/sidecar.go
@@ -1,0 +1,4 @@
+package cmd
+
+// TODO: `testground sidecar` command. Starts the sidecar daemon. To be
+// scheduled as a service in all cluster machines.
