+++ go.mod
@@ -14,43 +14,34 @@ require (
 	github.com/docker/docker v1.4.2-0.20190910181529-415f8ecb65e8
 	github.com/docker/go-connections v0.4.0 // indirect
 	github.com/docker/go-units v0.4.0 // indirect
-	github.com/dustin/go-humanize v1.0.0
-	github.com/go-redis/redis v6.15.5+incompatible
+	github.com/gogo/protobuf v1.3.0 // indirect
 	github.com/google/uuid v1.1.1
 	github.com/gorilla/mux v1.7.3 // indirect
+	github.com/hashicorp/go-cleanhttp v0.5.1 // indirect
 	github.com/hashicorp/go-getter v1.4.0
-	github.com/hashicorp/go-multierror v1.0.0
-	github.com/hashicorp/nomad/api v0.0.0-20190911182721-232295f83fc4
-	github.com/ipfs/go-datastore v0.1.0
-	github.com/ipfs/go-ipfs-api v0.0.2
-	github.com/ipfs/go-ipfs-config v0.0.11
-	github.com/ipfs/iptb v1.4.0
-	github.com/ipfs/iptb-plugins v0.2.0
 	github.com/ipfs/testground/sdk/runtime v0.0.0-00010101000000-000000000000
-	github.com/ipfs/testground/sdk/sync v0.0.0-00010101000000-000000000000 // indirect
-	github.com/libp2p/go-libp2p v0.3.1
-	github.com/libp2p/go-libp2p-core v0.2.2
-	github.com/libp2p/go-libp2p-kad-dht v0.2.1
+	github.com/mitchellh/go-homedir v1.1.0 // indirect
 	github.com/morikuni/aec v0.0.0-20170113033406-39771216ff4c // indirect
-	github.com/multiformats/go-multiaddr v0.0.4
 	github.com/opencontainers/go-digest v1.0.0-rc1 // indirect
 	github.com/opencontainers/image-spec v1.0.1 // indirect
 	github.com/opencontainers/runc v0.1.1 // indirect
 	github.com/otiai10/copy v1.0.1
 	github.com/otiai10/curr v0.0.0-20190513014714-f5a3d24e5776 // indirect
 	github.com/shurcooL/httpfs v0.0.0-20190707220628-8d4bc4ba7749 // indirect
-	github.com/shurcooL/httpgzip v0.0.0-20190720172056-320755c1c1b0 // indirect
 	github.com/shurcooL/vfsgen v0.0.0-20181202132449-6a9ea43bcacd
 	github.com/sirupsen/logrus v1.4.2 // indirect
+	github.com/stretchr/testify v1.3.0 // indirect
 	github.com/urfave/cli v1.22.0
+	go.opencensus.io v0.22.1 // indirect
 	go.uber.org/zap v1.10.0
 	golang.org/x/sys v0.0.0-20190910064555-bbd175535a8b // indirect
-	golang.org/x/tools/gopls v0.1.5 // indirect
+	golang.org/x/tools v0.0.0-20190911225940-c7d52e45e2f2 // indirect
 	google.golang.org/grpc v1.23.0 // indirect
 	gotest.tools v2.2.0+incompatible // indirect
 )
 
 replace (
+	github.com/ipfs/testground/sdk/iptb => ./sdk/iptb
 	github.com/ipfs/testground/sdk/runtime => ./sdk/runtime
 	github.com/ipfs/testground/sdk/sync => ./sdk/sync
 	github.com/miekg/dns => github.com/miekg/dns v1.0.14
