+++ sdk/README.md
@@ -1,2 +1,2 @@
-This package contains the software development kit for developing testground
-test plans in Go.
+This package contains the software development kit for developing test plans in
+Go.
