+++ sdk/runtime/doc.go
@@ -1,0 +1,3 @@
+// Package runtime contains functions and types to interact with the test
+// runtime environment, as formally defined in the system specification.
+package runtime
